# docs/specs/api_contract.yaml
openapi: 3.1.0
info:
  title: Derived API Contract
  version: 1.0.0
paths:
...
  /api/subsidiaries:
    get:
      operationId: get_api_subsidiaries
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: post_api_subsidiaries
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /api/subsidiaries/{id}:
    delete:
      operationId: delete_api_subsidiaries_id
      responses:
        '204':
          description: Derived from SPEC.md
    get:
      operationId: get_api_subsidiaries_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
    put:
      operationId: put_api_subsidiaries_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...